feat: started TODO in earnest
* Changed lib organization around data storage. Models go in lib/models
  and data storage implementations go in lib/data.
* Stopped tracking the database in git..
* Cleaned up the collections and data types and implementations.
* Updated some of the migrations to be more correct for the intended
  usage. Specifically, recipe revisions reference material_analysis, as
  simply referencing a material is too vague. All recipes are intended
  to have a full chemical analysis.
* Tested an oidc client against pocket-id. Works.

Still to do. A nice to have once the code is written is a db up command
defined in flake.nix. It should run `dbmate up` and a local script that
exercises the intended import commands through the api. This should
serve as both a "Get Started Quick" path and a sort of smoke test for
the core apis.